+++ MaxMind.GeoIP2/Model/NamedEntity.cs
@@ -1,9 +1,10 @@
 ﻿#region
 
 using MaxMind.Db;
-using Newtonsoft.Json;
 using System.Collections.Generic;
+using System.Collections.ObjectModel;
 using System.Linq;
+using System.Text.Json.Serialization;
 
 #endregion
 
@@ -14,9 +15,6 @@ namespace MaxMind.GeoIP2.Model
     /// </summary>
     public abstract class NamedEntity
     {
-        [JsonProperty("names")]
-        private readonly IDictionary<string, string> _names;
-
         /// <summary>
         ///     Constructor
         /// </summary>
@@ -24,10 +22,12 @@ namespace MaxMind.GeoIP2.Model
         protected NamedEntity(long? geoNameId = null, IDictionary<string, string>? names = null,
             IEnumerable<string>? locales = null)
         {
-            _names = names != null ? new Dictionary<string, string>(names) : new Dictionary<string, string>();
+            Names = new ReadOnlyDictionary<string, string>( 
+                names != null ? new Dictionary<string, string>(names) : new Dictionary<string, string>());
             // Unfortunately the existing models incorrectly use an int rather
             // than a long for the geoname_id. This should be corrected if we
             // ever do a major version bump.
+            // XXX - don't merge without fixing
             GeoNameId = (int?)geoNameId;
             Locales = locales != null ? new List<string>(locales) : new List<string> { "en" };
         }
@@ -40,13 +40,15 @@ namespace MaxMind.GeoIP2.Model
         ///         cred="GeoNameId" />
         ///     or relevant code instead.
         /// </summary>
-        [JsonIgnore]
-        public Dictionary<string, string> Names => new Dictionary<string, string>(_names);
+        [JsonInclude]
+        [JsonPropertyName("names")]
+        public IReadOnlyDictionary<string, string> Names { get; internal set; }
 
         /// <summary>
         ///     The GeoName ID for the city.
         /// </summary>
-        [JsonProperty("geoname_id")]
+        [JsonInclude]
+        [JsonPropertyName("geoname_id")]
         public int? GeoNameId { get; internal set; }
 
         /// <summary>
@@ -68,9 +70,8 @@ namespace MaxMind.GeoIP2.Model
         {
             get
             {
-                var names = _names;
-                var locale = Locales.FirstOrDefault(l => names.ContainsKey(l));
-                return locale == null ? null : names[locale];
+                var locale = Locales.FirstOrDefault(l => Names.ContainsKey(l));
+                return locale == null ? null : Names[locale];
             }
         }